AG-12625 - Align ApiReference styles with grid
