+++ external/ag-website-shared/src/utils/navigation.ts
@@ -1,6 +1,7 @@
 import type { Location, To, Update } from 'history';
 import { createBrowserHistory } from 'history';
-import { useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import type { MouseEventHandler } from 'react';
 
 export const browserHistory = import.meta.env.SSR ? null : createBrowserHistory();
 
@@ -12,6 +13,19 @@ export function useLocation() {
     const [location, setLocation] = useState<Location | null>(browserHistory?.location ?? null);
     useHistory(({ location }) => setLocation(location));
     return location;
+}
+
+export function useScrollToAnchor() {
+    const handleScroll: MouseEventHandler<HTMLAnchorElement> = useCallback((event) => {
+        event.preventDefault();
+        navigate(event.currentTarget.href);
+        const href = event.currentTarget.getAttribute('href');
+        if (href?.startsWith('#')) {
+            scrollIntoViewById(href.substring(1));
+        }
+    }, []);
+
+    return handleScroll;
 }
 
 export function navigate(to: To, options?: { state?: any; replace?: boolean }) {
